Working on auth, reorganized some files, using sjcl for hashing
