--- IQSignSENSE/backend/server.js
@@ -1,8 +1,0 @@
-
-import app from './app';
-
-const PORT = process.env.PORT || 3000;
-
-app.listen(PORT, () => {
-    console.log(`Server is running on http://localhost:${PORT}`);
-});
