+++ IQSignSENSE/eslint.config.js
@@ -36,16 +36,5 @@ export default [
     },
   },
 
-  // Configuration for Node.js (backend) files
-  {
-    files: ['backend/**/*.js'],
-    languageOptions: {
-      ecmaVersion: 2020,
-      globals: globals.node,
-      parserOptions: {
-        ecmaVersion: 'latest',
-        sourceType: 'module',
-      },
-    },
-  },
+
 ];
