+++ IQSignSENSE/src/App.jsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
-import RegistrationPage from './RegistrationPage';
-import LoginPage from './LoginPage';
+import RegistrationPage from './Auth/RegistrationPage';
+//import LoginPage from './LoginPage';
+import LoginPage from '../../iqsign/LoginPage';
 import ProfilePage from './ProfilePage';
 import Home from './HomePage';
 import SignEditor  from './Sign';
@@ -10,7 +11,7 @@ const App = () => {
   return (
     <Router>
       <Routes>
-        <Route path="/" element={<RegistrationPage />} />
+        <Route path="/register" element={<RegistrationPage />} />
         <Route path="/login" element={<LoginPage />} />
         <Route path="/profile" element={<ProfilePage />} />
         <Route path="/edit" element={<SignEditor />} />
