+++ iqsign/web/iqsign.js
@@ -708,6 +708,6 @@ function handleCodeGenerated(sts) {
    }
 }
 
-
+export default hasher;
 /* end of iqsign.js */
